requirements.txt
mcp[cli]>=1.0
asyncpg
psycopg2-binary
python-dotenv
python-telegram-bot>=21.0
anthropic
weasyprint
jinja2
